refactor(ui-next): share formatCpu, formatMemory, phaseKind and withRowAffordances
formatCpu, formatMemory, podStatusKind and withRowAffordances were each
duplicated between lib/kinds/columns.tsx, Resources.tsx and
Workloads.tsx — copied rather than exported because of the file-
ownership boundaries the parallel dispatches that wrote them drew.

Export the formatters and phaseKind (podStatusKind's superset) from
columns.tsx, and pull withRowAffordances into a new
lib/kinds/rowAffordances.tsx shared by both screens, generalised over
an isFlagged callback so it still serves a descriptor's flagged
function on one screen and a precomputed row.flagged boolean on the
other. No behaviour change: the full suite passes unchanged.
